finish with routing tests and implementation of all special cases
