+++ documentation/docs/routing-mechanism/routing.md
@@ -3,3 +3,32 @@ sidebar_position: 1
 ---
 # Intro
 
+```jsx title="/src/components/HelloCodeTitle.js"
+function HelloCodeTitle(props) {
+    // highlight-next-line
+  return <h1>Hello, {props.name}</h1>;
+}
+```
+
+```jsx live
+function Clock(props) {
+  const [date, setDate] = useState(new Date());
+  useEffect(() => {
+    var timerID = setInterval(() => tick(), 1000);
+
+    return function cleanup() {
+      clearInterval(timerID);
+    };
+  });
+
+  function tick() {
+    setDate(new Date());
+  }
+
+  return (
+    <div>
+      <h2>It is {date.toLocaleTimeString()}.</h2>
+    </div>
+  );
+}
+```
