+++ package.json
@@ -24,10 +24,10 @@
     "husky": "^7.0.4",
     "jest": "^27.5.1",
     "prettier": "^2.6.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-query": "^3.22.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.1",
     "rollup": "^2.72.0",
     "rollup-plugin-dts": "^4.2.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
